finsync-reality-lens/Backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
pillow
duckduckgo-search
transformers
torch